Got sel4zircon working on mq! Also qemu is bugged
Added config which could be used for mq benchmarking. Qemu constantly
firing irqs for hpet seems to be independent of the virtualised hpet
implementation? The callback to trigger irq in code is only called
when it's supposed to be. As such absolute timeouts should work, we
just need to account for extra irqs - need to sanity check any irqs
received!